+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Research pipeline.
+- Evaluation command for hovernet.
 
 ## [0.4.1] - 2023-03-28